+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.1.0
 
 * show node version on nodes overview, fixes [#260](https://github.com/cars10/elasticvue/issues/260)
+* add vim mode in editor, thanks @fqdeng, fixes [#253](https://github.com/cars10/elasticvue/pull/253)
 * update to vue `3.5`
 
 ## 1.0.11